// frontend/src/adapter/api/__generated/api.ts
export interface PostItemsRequestInner {
    /**
     * 
     * @type {string}
     * @memberof PostItemsRequestInner
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof PostItemsRequestInner
     */
    'description'?: string;
}